+++ RecPlayKontroller/rProjektListe.m
@@ -639,14 +639,14 @@ vomStart=derStatus;
 #pragma mark -
 #pragma mark ProjectTable Data Source:
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (long)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return [ProjektArray count];
 }
 
 - (id)tableView:(NSTableView *)aTableView
     objectValueForTableColumn:(NSTableColumn *)aTableColumn 
-			row:(int)rowIndex
+			row:(long)rowIndex
 {
     NSDictionary *einProjektDic;
 	//if (rowIndex<[ProjektArray count])
