+++ RecPlayKontroller/rProjektStart.m
@@ -303,14 +303,14 @@ extern const int StartmitDialog;//=2;
 #pragma mark ProjectTable Data Source:
 
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (long)numberOfRowsInTableView:(NSTableView *)aTableView
 {
    return [ProjektArray count];
 }
 
 - (id)tableView:(NSTableView *)aTableView
 objectValueForTableColumn:(NSTableColumn *)aTableColumn
-            row:(int)rowIndex
+            row:(long)rowIndex
 {
    NSDictionary *einProjektDic;
 	if (rowIndex<[ProjektArray count])
